pyproject.toml
[project]
name = "stock-monitoring"
version = "0.1.0"
description = "Vibe Coding Stock Monitoring System"
authors = [
    {name = "Antigravity Agent"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "numpy (>=2.4.0,<3.0.0)",
    "ccxt (>=4.5.30,<5.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "duckdb (>=1.4.3,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "feedparser (>=6.0.12,<7.0.0)",
    "asyncpg (>=0.29.0,<1.0.0)",
    "yfinance (>=0.2.36,<0.3.0)",
    "finance-datareader (>=0.9.60,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]
